fix: fix footer, add footer logo svg and refactor main page
