+++ src/components/index.ts
@@ -1,2 +1,2 @@
-export { Footer } from './Footer';
+export { Footer } from './Footer/Footer';
 export { Header } from './Header/Header';
