+++ src/pages/MainPage/MainPage.scss
@@ -26,6 +26,8 @@
 .main-create-session {
   display: flex;
   flex-direction: row;
+  justify-content: start;
+  align-items: center;
 }
 
 .main-connect-to-lobby {
