+++ src/pages/Start.tsx
@@ -3,7 +3,6 @@ import MainPage from './MainPage/MainPage';
 
 export const Start: React.FC = () => <div>
   <MainPage />
-  Start
 </div>;
 
 export default Start;
